机器学习与知识发现/Project/GPT-2/requirements.txt
absl-py==0.8.1
asn1crypto==0.24.0
astor==0.8.0
astropy==3.2.1
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
bokeh==1.2.0
boto3==1.10.8
botocore==1.13.8
Bottleneck==1.2.1
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
colorama==0.4.1
cPython==0.0.5
cryptography==2.7
cycler==0.10.0
cytoolz==0.9.0.1
dask==2.1.0
decorator==4.4.0
defusedxml==0.6.0
dill==0.2.9
distributed==2.1.0
docutils==0.15.2
entrypoints==0.3
fast-histogram==0.7
gast==0.2.2
glue-core==0.14.2
glue-vispy-viewers==0.11
google-auth==1.7.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.9.0
heapdict==1.0.0
idna==2.8
imageio==2.5.0
imgaug==0.3.0
importlib-metadata==0.17
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.14.0
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.5.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterthemes==0.20.0
Keras==2.0.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lesscpy==0.13.0
locket==0.2.0
lxml==4.4.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
more-itertools==7.0.0
mpl-scatter-density==0.5
msgpack==0.6.1
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.5
notebook==5.7.8
numpy==1.16.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.1.1.26
opencv-python-headless==4.1.1.26
opt-einsum==3.1.0
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.0
partd==1.0.0
pickleshare==0.7.5
Pillow==5.4.1
plotly==3.10.0
pluggy==0.12.0
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.10.0
psutil==5.6.3
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
Pygments==2.4.2
pymongo==3.9.0
PyOpenGL==3.1.1a1
pyOpenSSL==19.0.0
pyparsing==2.4.0
pypinyin==0.36.0
pyreadline==2.1
pyrsistent==0.14.11
PySocks==1.7.0
pytest==5.0.0
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.8.0
pytorch-pretrained-bert==0.6.2
pytz==2019.1
PyWavelets==1.0.3
pywinpty==0.5.5
PyYAML==5.1.1
pyzmq==18.0.0
qtconsole==4.5.1
QtPy==1.8.0
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.2.0
retrying==1.3.3
rsa==4.0
s3transfer==0.2.1
sacremoses==0.0.35
scikit-image==0.15.0
scipy==1.2.1
Send2Trash==1.5.0
sentencepiece==0.1.83
Shapely==1.6.4.post2
six==1.12.0
sortedcontainers==2.1.0
tblib==1.4.0
tensorboard==2.0.1
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
thulac==0.2.1
toolz==0.9.0
torch==1.0.1
torchtext==0.4.0
torchvision==0.2.1
tornado==6.0.3
tqdm==4.37.0
traitlets==4.3.2
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
zict==1.0.0
zipp==0.5.1
